Last Drop Server Operation App Foundation
This app requires a local SQL Server database to be initialized on the
local machine. Check out the /Database creation and population queries
to initialize a test database.

After the database is initialized you need to modify the  connection
string.

This app is using Alexa's store-from-database-to-application
functionality.

No operations are implemented yet.